# openapi.json
{"openapi":"3.1.0","info":{"title":"CareSense research prototype","description":"Synthetic ICU simulation with model-estimated risk. Not a diagnostic device. Use one local worker; simulation data resets on restart. Prototype thresholds are not clinically validated.","version":"0.1.0"},"paths":{"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/patients":{"get":{"summary":"Patients","operationId":"patients_patients_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}},"post":{"summary":"Add Patient","operationId":"add_patient_patients_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PatientInput"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/patients/{patient_id}":{"delete":{"summary":"Delete Patient","operationId":"delete_patient_patients__patient_id__delete","parameters":[{"name":"patient_id","in":"path","required":true,"schema":{"type":"string","title":"Patient Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/reset":{"post":{"summary":"Reset","operationId":"reset_reset_post","requestBody":{"content":{"application/json":{"schema":{"anyOf":[{"$ref":"#/components/schemas/EmptyInput"},{"type":"null"}],"title":"Body"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/vitals":{"get":{"summary":"Vitals","operationId":"vitals_vitals_get","parameters":[{"name":"patient_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string","minLength":1,"maxLength":128},{"type":"null"}],"title":"Patient Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"summary":"Change Vitals","operationId":"change_vitals_vitals_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VitalsInput"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/patients/{patient_id}/features":{"get":{"summary":"Features","operationId":"features_patients__patient_id__features_get","parameters":[{"name":"patient_id","in":"path","required":true,"schema":{"type":"string","title":"Patient Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/patients/{patient_id}/predict":{"post":{"summary":"Prediction","operationId":"prediction_patients__patient_id__predict_post","parameters":[{"name":"patient_id","in":"path","required":true,"schema":{"type":"string","title":"Patient Id"}},{"name":"top_n","in":"query","required":false,"schema":{"type":"integer","maximum":176,"minimum":1,"default":10,"title":"Top N"}}],"requestBody":{"content":{"application/json":{"schema":{"anyOf":[{"$ref":"#/components/schemas/EmptyInput"},{"type":"null"}],"title":"Body"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/patients/{patient_id}/explain":{"get":{"summary":"Explanation","operationId":"explanation_patients__patient_id__explain_get","parameters":[{"name":"patient_id","in":"path","required":true,"schema":{"type":"string","title":"Patient Id"}},{"name":"top_n","in":"query","required":false,"schema":{"type":"integer","maximum":176,"minimum":1,"default":10,"title":"Top N"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/patients/{patient_id}/trajectory":{"get":{"summary":"Trajectory","operationId":"trajectory_patients__patient_id__trajectory_get","parameters":[{"name":"patient_id","in":"path","required":true,"schema":{"type":"string","title":"Patient Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/alert-status/{patient_id}":{"get":{"summary":"Alert Status","operationId":"alert_status_alert_status__patient_id__get","parameters":[{"name":"patient_id","in":"path","required":true,"schema":{"type":"string","title":"Patient Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"EmptyInput":{"properties":{},"additionalProperties":false,"type":"object","title":"EmptyInput"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"PatientInput":{"properties":{"name":{"type":"string","maxLength":120,"minLength":1,"title":"Name"},"history":{"items":{"additionalProperties":true,"type":"object"},"type":"array","maxItems":10000,"minItems":1,"title":"History"}},"additionalProperties":false,"type":"object","required":["name","history"],"title":"PatientInput"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"VitalsInput":{"properties":{"patient_id":{"type":"string","maxLength":128,"minLength":1,"title":"Patient Id"},"expected_history_sha256":{"anyOf":[{"type":"string","pattern":"^[0-9a-f]{64}$"},{"type":"null"}],"title":"Expected History Sha256"},"history":{"anyOf":[{"items":{"additionalProperties":true,"type":"object"},"type":"array","maxItems":10000,"minItems":1},{"type":"null"}],"title":"History"},"observation":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Observation"}},"additionalProperties":false,"type":"object","required":["patient_id"],"title":"VitalsInput"}}}}